Move HTTP "Doers" to http package
