+++ internal/clients/http/proxy_doer.go
@@ -1,4 +1,4 @@
-package clients
+package http
 
 import (
 	"context"
